refactor: migrate hooks (use-timeline-chart) from web/app/ce to web/app/core
