+++ apps/web/core/hooks/use-timeline-chart.ts
@@ -6,13 +6,34 @@
 
 import { useContext } from "react";
 // types
+import type { TTimelineTypeCore } from "@plane/types";
 import type { TTimelineType } from "@plane/types";
+import { GANTT_TIMELINE_TYPE } from "@plane/types";
 // lib
 import { StoreContext } from "@/lib/store-context";
-// Plane-web
-import { getTimelineStore } from "@/plane-web/hooks/use-timeline-chart";
+// plane-web store types
 import type { IBaseTimelineStore } from "@/plane-web/store/timeline/base-timeline.store";
+import type { ITimelineStore } from "@/plane-web/store/timeline";
 import { useTimeLineType } from "../components/gantt-chart/contexts";
+
+export const getTimelineStore = (
+  timelineStore: ITimelineStore,
+  timelineType: TTimelineTypeCore
+): IBaseTimelineStore => {
+  if (timelineType === GANTT_TIMELINE_TYPE.ISSUE) {
+    return timelineStore.issuesTimeLineStore as IBaseTimelineStore;
+  }
+  if (timelineType === GANTT_TIMELINE_TYPE.MODULE) {
+    return timelineStore.modulesTimeLineStore as IBaseTimelineStore;
+  }
+  if (timelineType === GANTT_TIMELINE_TYPE.PROJECT) {
+    return timelineStore.projectTimeLineStore;
+  }
+  if (timelineType === GANTT_TIMELINE_TYPE.GROUPED) {
+    return timelineStore.groupedTimeLineStore;
+  }
+  throw new Error(`Unknown timeline type: ${timelineType}`);
+};
 
 export const useTimeLineChart = (timelineType: TTimelineType): IBaseTimelineStore => {
   const context = useContext(StoreContext);
